+++ backend/src/services/utils/resampling.py
@@ -5,7 +5,7 @@ import numpy as np
 import pyarrow as pa
 
 from src.services.utils.field_metadata import is_rate_from_field_meta
-from src.services.sumo_access.types.summary_types import Frequency
+from src.services.types.summary_types import Frequency
 
 
 def _truncate_day_to_monday(datetime_day: np.datetime64) -> np.datetime64:
